requirements.txt
fastapi
requests
python-dotenv
selenium